+++ src/dipdup/index.py
@@ -605,8 +605,10 @@ class OperationIndex(Index):
 
     async def _get_origination_addresses(self) -> Set[str]:
         """Get addresses to fetch origination from during initial synchronization"""
-        # FIXME: Missing `OperationType.origination` in config is ignored
-        addresses = set()
+        if OperationType.origination not in self._config.types:
+            return set()
+
+        addresses: Set[str] = set()
         for handler_config in self._config.handlers:
             for pattern_config in handler_config.pattern:
                 if not isinstance(pattern_config, OperationHandlerOriginationPatternConfig):